# docs/swagger.yaml
definitions:
  jsonpatch.Operation:
    additionalProperties:
      items:
        type: integer
      type: array
    type: object
...
  pb.GeodeticPosition:
    properties:
      altitude:
        maximum: 99999
        minimum: 0
        type: number
      latitude:
        maximum: 90
        minimum: -90
        type: number
      longitude:
        maximum: 180
        minimum: -180
        type: number
    type: object
...
  pb.Location:
    properties:
      geodetic_position:
        allOf:
        - $ref: '#/definitions/pb.GeodeticPosition'
        description: string  ID          = 1;//`json:"id,omitempty" bson:"_id"`
    type: object
...
  pb.PatchResponse:
    properties:
      IsOk:
        type: boolean
      Message:
        type: string
    type: object
...
  pb.TrackHistory:
    properties:
      created_at:
        type: integer
      datasource:
        type: string
      drone_id:
        type: string
      id:
        type: string
      location_byte:
        items:
          type: integer
        type: array
      order_id:
        type: string
      track_id:
        type: integer
    type: object
...
  types.ErrorResponse:
    properties:
      code:
        example: 400
        type: integer
      message:
        example: status bad request
        type: string
    type: object
  util.TrackHistoryByteAndJson:
    properties:
      location_json:
        $ref: '#/definitions/pb.Location'
      track_history:
        $ref: '#/definitions/pb.TrackHistory'
    type: object
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a sample Go Template server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Swagger Go Template Project API
  version: "1.0"
paths:
...
  /track_historys:
    get:
      consumes:
      - application/json
      description: Find all track_history
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.TrackHistoryByteAndJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Find all track_history
      tags:
      - track_historys
    post:
      consumes:
      - application/json
      description: Create a new track_history
      parameters:
      - description: track_history body
        in: body
        name: track_history
        required: true
        schema:
          $ref: '#/definitions/pb.TrackHistory'
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      - description: event api call flag
        in: query
        name: lat
        required: true
        type: number
      - description: event api call flag
        in: query
        name: lon
        required: true
        type: number
      - description: event api call flag
        in: query
        name: alt
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.TrackHistoryByteAndJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Create  track_history
      tags:
      - track_historys
  /track_historys/{id}:
    delete:
      consumes:
      - application/json
      description: Delete track_history by ID
      parameters:
      - description: track_history id
        in: path
        name: id
        required: true
        type: string
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.TrackHistoryByteAndJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Delete track_history by ID
      tags:
      - track_historys
    get:
      consumes:
      - application/json
      description: Find track_history by ID
      parameters:
      - description: drone id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.TrackHistoryByteAndJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Find track_history by ID
      tags:
      - track_historys
    patch:
      consumes:
      - application/json
      description: Patch track_history by ID use standard of JSON PATCH https://jsonpatch.com/
      parameters:
      - description: track_history id
        in: path
        name: id
        required: true
        type: string
      - description: Patch operation format Array of Operation Add, Remove, Replace,
          Copy, Move, Test. Get example at https://jsonpatch.com/
        in: body
        name: drone
        required: true
        schema:
          items:
            $ref: '#/definitions/jsonpatch.Operation'
          type: array
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/pb.PatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Patch track_history by ID
      tags:
      - track_historys
    put:
      consumes:
      - application/json
      description: Update track_history by ID
      parameters:
      - description: track_history id
        in: path
        name: id
        required: true
        type: string
      - description: track_history body
        in: body
        name: track_history
        required: true
        schema:
          $ref: '#/definitions/pb.TrackHistory'
      - description: event api call flag
        in: query
        name: eventAPI
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.TrackHistoryByteAndJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Update track_history by ID
      tags:
      - track_historys
  /track_historys/search:
    get:
      consumes:
      - application/json
      description: Search track_history use Query option https://github.com/jtlabsio/mongo/
      parameters:
      - description: page number
        in: query
        name: page[page]
        required: true
        type: integer
      - description: page size
        in: query
        name: page[size]
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/util.TrackHistoryByteAndJson'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/types.ErrorResponse'
      summary: Search track_history
      tags:
      - track_historys
swagger: "2.0"